refactor: 회원가입 로직 user -> auth 컨트롤러로 이동
# Conflicts:
#	user/src/main/java/com/homeaid/repository/UserRepository.java
