+++ user/src/main/java/com/homeaid/repository/UserRepository.java
@@ -1,7 +1,6 @@
 package com.homeaid.repository;
 
 import com.homeaid.domain.User;
-import java.time.LocalDateTime;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
